Refactor validate! methods out of initialize
Previously, the validate! method was called from the corresponding
constructor, which is handy because it prevents objects existing with
invalid data.

However, it makes it hard to test, as the data always needs to be
valid. This change makes the validate! method need to be manually
called, and the call is manually added everywhere the gem creates
instances of the relevant objects.